仓租新增库位计算规则
(cherry picked from commit 648587981e114112ffb80c1e5d1ee2af1bed2509)

+++ business-center/wms-center/wms-service/wms-business-chargerules/src/main/java/com/szmsd/chargerules/service/impl/WarehouseOperationServiceImpl.java
@@ -9,6 +9,7 @@ import com.szmsd.bas.vo.BasSellerInfoVO;
 import com.szmsd.chargerules.domain.WarehouseOperation;
 import com.szmsd.chargerules.domain.WarehouseOperationDetails;
 import com.szmsd.chargerules.dto.WarehouseOperationDTO;
+import com.szmsd.chargerules.mapper.WarehouseOperationDetailsMapper;
 import com.szmsd.chargerules.mapper.WarehouseOperationMapper;
 import com.szmsd.chargerules.service.IWarehouseOperationDetailsService;
 import com.szmsd.chargerules.service.IWarehouseOperationService;
@@ -21,6 +22,7 @@ import com.szmsd.common.core.utils.bean.BeanMapperUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.beans.BeanUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -45,6 +47,8 @@ public class WarehouseOperationServiceImpl extends ServiceImpl<WarehouseOperatio
     private IWarehouseOperationDetailsService warehouseOperationDetailsService;
     @Resource
     private BasSellerFeignService basSellerFeignService;
+    @Autowired
+    private WarehouseOperationDetailsMapper warehouseOperationDetailsMapper;
 
     private final String REGEX = "\\d+D\\d+";
 
@@ -54,14 +58,27 @@ public class WarehouseOperationServiceImpl extends ServiceImpl<WarehouseOperatio
         if (!checkWarehouse(dto)) {
             WarehouseOperation domain = new WarehouseOperation();
             BeanUtils.copyProperties(dto, domain);
-            AssertUtil.notEmpty(dto.getDetails(), "详情必填");
+
             warehouseOperationMapper.insert(domain);
-            long count = dto.getDetails().stream().peek(value -> value.setWarehouseOperationId(domain.getId())).filter(value -> !Pattern.matches(REGEX, value.getChargeDays())).count();
-            if (count > 0) throw new CommonException("999", "计费天数格式错误");
+            if (dto.getDetails()!=null){
+                AssertUtil.notEmpty(dto.getDetails(), "详情必填");
+                long count = dto.getDetails().stream().peek(value -> value.setWarehouseOperationId(domain.getId())).filter(value -> !Pattern.matches(REGEX, value.getChargeDays())).count();
+                if (count > 0) throw new CommonException("999", "计费天数格式错误");
+
+            }
             List<WarehouseOperationVo> warehouseOperationDb = this.listPage(dto);
-            if (isIntersection(dto.getDetails(), warehouseOperationDb.get(0).getDetails()))
-                throw new CommonException("999", "仓库+区间存在重合");
-            warehouseOperationDetailsService.saveBatch(dto.getDetails());
+            if ( dto.getDetails()!=null) {
+                if (isIntersection(dto.getDetails(), warehouseOperationDb.get(0).getDetails()))
+                    throw new CommonException("999", "仓库+区间存在重合");
+            }
+
+            if (dto.getDetails()!=null) {
+                warehouseOperationDetailsService.saveBatch(dto.getDetails());
+            }
+            if (dto.getLocationDetails()!=null) {
+                AssertUtil.notEmpty(dto.getLocationDetails(), "详情必填");
+                warehouseOperationDetailsService.saveBatch(dto.getLocationDetails());
+            }
         }
         return 1;
     }
@@ -74,12 +91,20 @@ public class WarehouseOperationServiceImpl extends ServiceImpl<WarehouseOperatio
      */
     private boolean checkWarehouse(WarehouseOperationDTO dto) {
         AssertUtil.notNull(dto.getWarehouseCode(), "仓库必填");
-        List<WarehouseOperationVo> list = this.listPage(dto);
-        if (CollectionUtils.isNotEmpty(list)) { //仓库已存在 添加到该仓库下面的详情
-            if (isIntersection(dto.getDetails(), list.get(0).getDetails()))
-                throw new CommonException("999", "仓库+区间存在重合");
-            List<WarehouseOperationDetails> collect = dto.getDetails().stream().peek(value -> value.setWarehouseOperationId(list.get(0).getId())).collect(Collectors.toList());
-            return warehouseOperationDetailsService.saveBatch(collect);
+
+            if (dto.getDetails()!=null) {
+                List<WarehouseOperationVo> list = this.listPage(dto);
+                if (CollectionUtils.isNotEmpty(list)) { //仓库已存在 添加到该仓库下面的详情
+                    if (isIntersection(dto.getDetails(), list.get(0).getDetails()))
+                        throw new CommonException("999", "仓库+区间存在重合");
+
+                    List<WarehouseOperationDetails> collect = dto.getDetails().stream().peek(value -> value.setWarehouseOperationId(list.get(0).getId())).collect(Collectors.toList());
+                return warehouseOperationDetailsService.saveBatch(collect);
+            }
+            if (dto.getLocationDetails()!=null) {
+                List<WarehouseOperationDetails> collect = dto.getLocationDetails().stream().peek(value -> value.setWarehouseOperationId(list.get(0).getId())).collect(Collectors.toList());
+                warehouseOperationDetailsService.saveBatch(collect);
+            }
         }
         return false;
     }
@@ -115,8 +140,10 @@ public class WarehouseOperationServiceImpl extends ServiceImpl<WarehouseOperatio
     @Transactional
     @Override
     public int update(WarehouseOperationDTO dto) {
-        long count = dto.getDetails().stream().peek(value -> value.setWarehouseOperationId(dto.getId())).filter(value -> !Pattern.matches(REGEX, value.getChargeDays())).count();
-        if (count > 0) throw new CommonException("999", "计费天数格式错误");
+        if (dto.getDetails()!=null){
+            long count = dto.getDetails().stream().peek(value -> value.setWarehouseOperationId(dto.getId())).filter(value -> !Pattern.matches(REGEX, value.getChargeDays())).count();
+            if (count > 0) throw new CommonException("999", "计费天数格式错误");
+        }
         WarehouseOperation map = BeanMapperUtil.map(dto, WarehouseOperation.class);
         this.updateDetails(dto);
         return warehouseOperationMapper.updateById(map);
@@ -125,9 +152,26 @@ public class WarehouseOperationServiceImpl extends ServiceImpl<WarehouseOperatio
     private void updateDetails(WarehouseOperationDTO dto) {
         LambdaQueryWrapper<WarehouseOperationDetails> query = Wrappers.lambdaQuery();
         query.eq(WarehouseOperationDetails::getWarehouseOperationId, dto.getId());
-        warehouseOperationDetailsService.remove(query);
-        if (isIntersection(dto.getDetails(), new ArrayList<>())) throw new CommonException("999", "仓库+区间存在重合");
-        warehouseOperationDetailsService.saveBatch(dto.getDetails());
+
+        if (dto.getLocationDetails()!=null){
+            dto.getLocationDetails().forEach(x->{
+                x.setWarehouseOperationId(dto.getId());
+            });
+            query.eq(WarehouseOperationDetails::getComputeType,1);
+            warehouseOperationDetailsService.remove(query);
+            warehouseOperationDetailsService.saveBatch(dto.getLocationDetails());
+
+        }
+        query.eq(WarehouseOperationDetails::getWarehouseOperationId, dto.getId());
+        if (dto.getDetails()!=null){
+            query.eq(WarehouseOperationDetails::getComputeType,0);
+            warehouseOperationDetailsService.remove(query);
+            if (isIntersection(dto.getDetails(), new ArrayList<>())) throw new CommonException("999", "仓库+区间存在重合");
+            warehouseOperationDetailsService.saveBatch(dto.getDetails());
+
+        }
+
+
     }
 
     @Override
@@ -185,7 +229,18 @@ public class WarehouseOperationServiceImpl extends ServiceImpl<WarehouseOperatio
 
     @Override
     public WarehouseOperationVo details(int id) {
-        return warehouseOperationMapper.selectDetailsById(id);
+        WarehouseOperationVo warehouseOperationVo=warehouseOperationMapper.selectDetailsById(id);
+        //库存
+        List<WarehouseOperationDetails> warehouseOperationDetails=warehouseOperationDetailsMapper.selectWarehouseOperationDetailsrs(id,0);
+        //库位
+        List<WarehouseOperationDetails> locationwarehouseOperationDetails=warehouseOperationDetailsMapper.selectWarehouseOperationDetailsrs(id,1);
+        if (warehouseOperationDetails.size()>0){
+            warehouseOperationVo.setDetails(warehouseOperationDetails);
+        }
+        if (locationwarehouseOperationDetails.size()>0){
+            warehouseOperationVo.setLocationDetails(locationwarehouseOperationDetails);
+        }
+        return warehouseOperationVo;
     }
 
 }
